+++ rest-dynamodb-modular/app/src/handlers/__tests__/item-find.test.js
@@ -1,10 +1,10 @@
-// Import all functions from delete-item.js
-const lambda = require('../../../handlers/item-delete');
+// Import all functions from find-item.js
+const lambda = require('../item-find');
 // Import dynamodb from aws-sdk
 const { DynamoDBDocumentClient } = require('@aws-sdk/lib-dynamodb');
 
 // This includes all tests for FindItem handler
-describe('handler::DeleteItem', () => {
+describe('handler::FindItem', () => {
   let sendSpy;
 
   // Test one-time setup and teardown, see more in https://jestjs.io/docs/en/setup-teardown
@@ -24,14 +24,14 @@ describe('handler::DeleteItem', () => {
     sendSpy.mockRestore();
   });
 
-  it('should return status code 204 when successful', async () => {
+  it('should return item when found', async () => {
     const item = { id: 'id1' };
 
     // Return the specified value whenever the spied function is called
-    sendSpy.mockResolvedValue({});
+    sendSpy.mockResolvedValue({ Item: item });
 
     const event = {
-      httpMethod: 'DELETE',
+      httpMethod: 'GET',
       pathParameters: {
         itemId: 'id1',
       },
@@ -41,7 +41,34 @@ describe('handler::DeleteItem', () => {
     const result = await lambda.handle(event);
 
     const expectedResult = {
-      statusCode: 204,
+      statusCode: 200,
+      body: JSON.stringify(item),
+    };
+
+    // Expect dynamodb to have been called
+    expect(sendSpy).toHaveBeenCalledTimes(1);
+    // Compare the result with the expected result
+    expect(result).toEqual(expectedResult);
+  });
+
+  it('should return status code 404 when not found', async () => {
+    const item = { id: 'id1' };
+
+    // Return the specified value whenever the spied function is called
+    sendSpy.mockReturnValue(Promise.resolve({}));
+
+    const event = {
+      httpMethod: 'GET',
+      pathParameters: {
+        itemId: 'id1',
+      },
+    };
+
+    // Invoke the handler
+    const result = await lambda.handle(event);
+
+    const expectedResult = {
+      statusCode: 404,
     };
 
     // Expect dynamodb to have been called
